refactor metrics let it initialize by configuration
