+++ metrics/seata-metrics-core/pom.xml
@@ -33,6 +33,11 @@
             <artifactId>seata-metrics-api</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>seata-config-core</artifactId>
+            <version>${project.version}</version>
+        </dependency>
     </dependencies>
 
 </project>
