--- metrics/seata-metrics-core/src/main/resources/META-INF/services/io.seata.metrics.Registry
@@ -1,1 +1,0 @@
-io.seata.metrics.DefaultRegistry
