+++ metrics/seata-metrics-exporter-prometheus/pom.xml
@@ -23,29 +23,25 @@
         <version>${revision}</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>seata-metrics-all</artifactId>
-    <name>seata-metrics-all ${project.version}</name>
+
+    <artifactId>seata-metrics-exporter-prometheus</artifactId>
+    <name>seata-metrics-exporter-prometheus ${project.version}</name>
+
+    <properties>
+        <prometheus.client.version>0.6.0</prometheus.client.version>
+    </properties>
 
     <dependencies>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>seata-metrics-api</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>${project.groupId}</groupId>
             <artifactId>seata-metrics-core</artifactId>
-            <version>${project.version}</version>
+            <version>${project.parent.version}</version>
         </dependency>
+
         <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>seata-metrics-registry-compact</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>seata-metrics-exporter-prometheus</artifactId>
-            <version>${project.version}</version>
+            <groupId>io.prometheus</groupId>
+            <artifactId>simpleclient_httpserver</artifactId>
+            <version>${prometheus.client.version}</version>
         </dependency>
     </dependencies>
