+++ metrics/seata-metrics-registry-compact/src/main/resources/META-INF/services/io.seata.metrics.registry.Registry
@@ -1,0 +1,1 @@
+io.seata.metrics.registry.compact.CompactRegistry
