+++ server/src/main/resources/file.conf
@@ -109,3 +109,11 @@ transaction {
   undo.data.validation = true
   undo.log.serialization = jackson
 }
+
+## metrics settings
+#metrics {
+#  registry-type = "compact"
+#  # multi exporters use comma divided
+#  exporter-list = "prometheus"
+#  exporter-prometheus-port = 9898
+#}
