+++ server/src/test/resources/file.conf
@@ -3,3 +3,11 @@ recovery {
   asyn-committing-retry-delay = 1
   timeout-retry-delay = 1
 }
+
+## metrics settings
+metrics {
+  registry-type = "compact"
+  # multi exporters use comma divided
+  exporter-list = "prometheus"
+  exporter-prometheus-port = 9898
+}
